Move ConnectWallet modal to its own context (#208)
* Initial Deposit/Withdraw view

 - Add BaseModal, refactor ConnectWallet modal to this later

* More progress

* Polish the UI

* Fix swap

* Better select market view and fix X close button

* Move ConnectWallet modal to its own context

* Delete unused imports

---------
